+++ CHANGELOG.rst
@@ -1,3 +1,8 @@
+0.2.0 (2016-04-28)
+-----------------
+* Tf lookup from transformation between marker's and frontcam's frame and provide it as next goal to be reached
+
+
 0.1.0 (2016-04-15)
 ------------------
-* Tf subscruber to /ar_pose_marker
+* Tf subscriber to /ar_pose_marker